+++ packages/terra-clinical-result/CHANGELOG.md
@@ -3,7 +3,9 @@ ChangeLog
 
 Unreleased
 -----------------
+### Fixed
+* Fix bug when ClinicalBloodPresure was provided both a systolic and diastolic no data, presented error
 
 1.0.0 - (March 6, 2020)
 ----------
-* Initial stable release 
+* Initial stable release