+++ packages/terra-clinical-result/CHANGELOG.md
@@ -3,7 +3,9 @@ ChangeLog
 
 Unreleased
 -----------------
+### Fixed
+* Fixed bug where ClinicalBloodPresure if provided both a systolic and diastolic object with `resultNoData` true, presented error result instead of no data result.
 
 1.0.0 - (March 6, 2020)
 ----------
-* Initial stable release 
+* Initial stable release