+++ packages/terra-clinical-result/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 -----------------
+1.1.0 - (March 10, 2020)
+----------
 ### Fixed
 * Fixed bug where ClinicalBloodPresure if provided both a systolic and diastolic object with `resultNoData` true, presented error result instead of no data result.